Add --with-colons option
This option tells gpg to use a colon-separated output format intended for parsing
by scripts rather than reading by humans. From my experience it is also less likely
to truncate the key id we are grepping for here, which would cause the grep
command to miss the key id, even when it exists in the key database.